move _conventional_xc function for a more consistent layout
